[bugfix] Tabs: incorrect line position in some cases (#3961)
