+++ src/sticky/test/demo.spec.js
@@ -1,6 +1,4 @@
 import Demo from '../demo';
 import demoTest from '../../../test/demo-test';
-import { mockHTMLElementOffset } from '../../../test/utils';
 
-mockHTMLElementOffset();
 demoTest(Demo);
