+++ src/utils/dom/style.ts
@@ -1,0 +1,5 @@
+export function isHidden(element: HTMLElement) {
+  return (
+    window.getComputedStyle(element).display === 'none' || element.offsetParent === null
+  );
+}
