+++ test/utils/transition.ts
@@ -1,0 +1,4 @@
+import Vue from 'vue';
+import { TransitionStub } from '@vue/test-utils';
+
+Vue.component('transition', TransitionStub as any);
